Add Quick Initialization IpMappings naming based on IP UltraParallel processing, using bundled tf files for resources and splitting modules Use SLOv2 instead of SLOv1 dependency for dashboards

+++ dynatrace/export/environment.go
@@ -18,6 +18,7 @@
 package export
 
 import (
+	"bufio"
 	"bytes"
 	"encoding/json"
 	"fmt"
@@ -29,6 +30,7 @@ import (
 	"strings"
 	"sync"
 
+	"github.com/dynatrace-oss/terraform-provider-dynatrace/dynatrace/address"
 	"github.com/dynatrace-oss/terraform-provider-dynatrace/dynatrace/api/v2/entity"
 	entitysettings "github.com/dynatrace-oss/terraform-provider-dynatrace/dynatrace/api/v2/entity/settings"
 	"github.com/dynatrace-oss/terraform-provider-dynatrace/dynatrace/settings"
@@ -40,6 +42,10 @@ import (
 )
 
 var NO_REFRESH_ON_IMPORT = os.Getenv("DYNATRACE_NO_REFRESH_ON_IMPORT") == "true"
+var QUICK_INIT = os.Getenv("DYNATRACE_QUICK_INIT") == "true"
+var ULTRA_PARALLEL = os.Getenv("DYNATRACE_ULTRA_PARALLEL") == "true"
+
+const ENV_VAR_CUSTOM_PROVIDER_LOCATION = "DYNATRACE_CUSTOM_PROVIDER_LOCATION"
 
 type Environment struct {
 	mu                    sync.Mutex
@@ -51,7 +57,10 @@ type Environment struct {
 	ChildResourceOverride bool
 	PrevStateMapCommon    *StateMap
 	PrevNamesByModule     map[string][]string
+	ImportStateMap        *StateMap
 	ChildParentGroups     map[ResourceType]ResourceType
+	IsParentMap           map[ResourceType]bool
+	HasDependenciesTo     map[ResourceType]bool
 }
 
 func (me *Environment) TenantID() string {
@@ -112,17 +121,63 @@ func (me *Environment) Export() (err error) {
 
 func (me *Environment) PreProcess() error {
 	me.ProcessChildParentGroups()
+	me.ProcessHasDependenciesTo()
+	err := me.LoadImportState()
+	if err != nil {
+		return err
+	}
+	err = me.ProcessPrevState()
+	if err != nil {
+		return err
+	}
 
-	return me.ProcessPrevState()
+	return nil
 }
+
 func (me *Environment) ProcessChildParentGroups() {
 	me.ChildParentGroups = map[ResourceType]ResourceType{}
+	me.IsParentMap = map[ResourceType]bool{}
+
 	for resType, resource := range AllResources {
 		if resource.Parent != nil {
 			me.ChildParentGroups[resType] = *resource.Parent
 			me.ChildParentGroups[*resource.Parent] = *resource.Parent
+			me.IsParentMap[*resource.Parent] = true
 		}
 	}
+}
+
+func (me *Environment) ProcessHasDependenciesTo() {
+	me.HasDependenciesTo = map[ResourceType]bool{}
+
+	for _, resource := range AllResources {
+		for _, dep := range resource.Dependencies {
+			resSource := dep.ResourceType()
+			if resSource == "" {
+				continue
+			}
+
+			me.HasDependenciesTo[resSource] = true
+
+		}
+	}
+}
+
+func (me *Environment) LoadImportState() error {
+	if IMPORT_STATE_PATH == "" {
+		return nil
+	}
+
+	state, err := LoadStateFile(IMPORT_STATE_PATH)
+	if err != nil {
+		return err
+	}
+
+	stateMap := BuildStateMap(state)
+
+	me.ImportStateMap = stateMap
+
+	return nil
 }
 
 func (me *Environment) ProcessPrevState() error {
@@ -326,6 +381,7 @@ func (me *Environment) PostProcess() error {
 		}
 	}
 
+	fmt.Println("Post-Processing Resources - Group child configs with parent configs ...")
 	for _, resource := range me.GetChildResources() {
 		if resource.GetParent().Status == ResourceStati.Erronous {
 			continue
@@ -364,6 +420,7 @@ func getResMap(resources []*Resource) map[ResourceType][]*Resource {
 }
 
 func (me *Environment) Finish() (err error) {
+	fmt.Println("Finishing touches ...")
 	if shutdown.System.Stopped() {
 		return nil
 	}
@@ -403,7 +460,7 @@ func (me *Environment) Module(resType ResourceType) *Module {
 		Status:               ModuleStati.Untouched,
 		Environment:          me,
 		ChildParentIDNameMap: map[string]string{},
-		ChildParentMutex:     new(sync.Mutex),
+		ModuleMutex:          new(sync.Mutex),
 	}
 	me.Modules[resType] = module
 	return module
@@ -571,6 +628,7 @@ func (me *Environment) WriteResourceFiles() (err error) {
 }
 
 func (me *Environment) RemoveNonReferencedModules() (err error) {
+	fmt.Println("Remove Non-Referenced Modules ...")
 	m := map[ResourceType]*Module{}
 	for k, module := range me.Modules {
 		m[k] = module
@@ -580,7 +638,9 @@ func (me *Environment) RemoveNonReferencedModules() (err error) {
 			module.PurgeFolder()
 			delete(me.Modules, k)
 		} else if !module.Environment.ChildResourceOverride && module.Descriptor.Parent != nil {
-			module.PurgeFolder()
+			if me.Flags.FollowReferences {
+				module.PurgeFolder()
+			}
 		} else if len(module.GetPostProcessedResources()) == 0 {
 			module.PurgeFolder()
 			delete(me.Modules, k)
@@ -590,9 +650,52 @@ func (me *Environment) RemoveNonReferencedModules() (err error) {
 }
 
 func (me *Environment) WriteProviderFiles() (err error) {
-	fmt.Println("Writing ___providers___.tf")
 
+	if QUICK_INIT {
+		// pass
+	} else {
+		err = me.WriteMainProviderFile()
+		if err != nil {
+			return err
+		}
+	}
+
+	if me.Flags.Flat {
+		return nil
+	}
+
+	fmt.Println("Writing modules ___providers___.tf")
+	parallel := (os.Getenv("DYNATRACE_PARALLEL") != "false")
+	if parallel {
+		var wg sync.WaitGroup
+		wg.Add(len(me.Modules))
+		for _, module := range me.Modules {
+			go func(module *Module) error {
+				defer wg.Done()
+				if shutdown.System.Stopped() {
+					return nil
+				}
+				if err = module.WriteProviderFile(false); err != nil {
+					return err
+				}
+				return nil
+			}(module)
+		}
+		wg.Wait()
+	} else {
+		for _, module := range me.Modules {
+			if err = module.WriteProviderFile(true); err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+func (me *Environment) WriteMainProviderFile() error {
+	fmt.Println("Writing main ___providers___.tf")
 	var outputFile *os.File
+	var err error = nil
 	if outputFile, err = me.CreateFile("___providers___.tf"); err != nil {
 		return err
 	}
@@ -623,34 +726,7 @@ func (me *Environment) WriteProviderFiles() (err error) {
 `, providerSource, providerVersion)); err != nil {
 		return err
 	}
-	if me.Flags.Flat {
-		return nil
-	}
 
-	parallel := (os.Getenv("DYNATRACE_PARALLEL") != "false")
-	if parallel {
-		var wg sync.WaitGroup
-		wg.Add(len(me.Modules))
-		for _, module := range me.Modules {
-			go func(module *Module) error {
-				defer wg.Done()
-				if shutdown.System.Stopped() {
-					return nil
-				}
-				if err = module.WriteProviderFile(false); err != nil {
-					return err
-				}
-				return nil
-			}(module)
-		}
-		wg.Wait()
-	} else {
-		for _, module := range me.Modules {
-			if err = module.WriteProviderFile(true); err != nil {
-				return err
-			}
-		}
-	}
 	return nil
 }
 
@@ -734,9 +810,9 @@ func (me *Environment) WriteMainFile() error {
 		if len(me.Module(resourceType).GetPostProcessedResources()) == 0 {
 			continue
 		}
-		mainFile.WriteString(fmt.Sprintf("module \"%s\" {\n", resourceType.Trim()))
+
 		module := me.Module(resourceType)
-		mainFile.WriteString(fmt.Sprintf("  source = \"./%s\"\n", module.GetFolder(true)))
+		me.writeOpeningMainSection(mainFile, resourceType.Trim(), module.GetFolder(true))
 
 		if ATOMIC_DEPENDENCIES {
 
@@ -773,14 +849,33 @@ func (me *Environment) WriteMainFile() error {
 				}
 			}
 		}
-		mainFile.WriteString("}\n\n")
+		writeClosingMainSection(mainFile)
+
+		if module.SplitPathModuleNameMap != nil {
+			for _, splitName := range module.SplitPathModuleNameMap {
+				me.writeOpeningMainSection(mainFile, splitName, fmt.Sprintf("./modules/%s", splitName))
+				writeClosingMainSection(mainFile)
+			}
+		}
 	}
 	return nil
 }
 
+func (me *Environment) writeOpeningMainSection(mainFile *os.File, trimmedResourceType string, resourceFolder string) {
+	mainFile.WriteString(fmt.Sprintf("module \"%s\" {\n", trimmedResourceType))
+	mainFile.WriteString(fmt.Sprintf("  source = \"./%s\"\n", resourceFolder))
+}
+
+func writeClosingMainSection(mainFile *os.File) {
+	mainFile.WriteString("}\n\n")
+}
+
 func (me *Environment) ExecuteImport() error {
 	if me.Flags.ImportStateV2 {
-		return me.executeImportV2()
+		fmt.Println("Importing Resources into Terraform State ...")
+		err := me.executeImportV2()
+		fmt.Println("Imported Resources into Terraform State ...")
+		return err
 	}
 
 	return nil
@@ -922,4 +1017,155 @@ func (me *Environment) executeTF(cmd *exec.Cmd) (err error) {
 	}
 
 	return nil
+}
+
+func (me *Environment) FinishExport() error {
+
+	fmt.Println("Finish Export ...")
+	address.SaveOriginalMap(me.OutputFolder)
+	address.SaveCompletedMap(me.OutputFolder)
+
+	if QUICK_INIT {
+		err := me.WriteQuickModulesJSON()
+		if err != nil {
+			return err
+		}
+	} else if me.Flags.SkipTerraformInit {
+		// pass
+	} else {
+		err := me.RunTerraformInit()
+		if err != nil {
+			return err
+		}
+	}
+
+	if me.Flags.ImportStateV2 {
+		if err := me.ExecuteImport(); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (me *Environment) RunTerraformInit() error {
+	exePath, _ := exec.LookPath("terraform")
+	fmt.Println("Terraform executable path: ", exePath)
+	cmdOptions := []string{"init", "-no-color"}
+
+	customProviderLocation := os.Getenv(ENV_VAR_CUSTOM_PROVIDER_LOCATION)
+	if len(customProviderLocation) != 0 && customProviderLocation != "" {
+		cmdOptions = append(cmdOptions, fmt.Sprint("-plugin-dir=", customProviderLocation))
+	}
+
+	cmd := exec.Command(exePath, cmdOptions...)
+	cmd.Dir = me.OutputFolder
+	outs, err := cmd.StderrPipe()
+	if err != nil {
+		return err
+	}
+	err = cmd.Start()
+	if err != nil {
+		fmt.Println("Terraform CLI not installed - skipping import")
+		return nil
+	} else {
+		fmt.Println("Executing 'terraform init'")
+		defer func() {
+			cmd.Wait()
+		}()
+
+		go readStuff(bufio.NewScanner(outs))
+	}
+	return nil
+}
+
+type TerraformInitModuleList struct {
+	Modules []TerraformInitModule `json:"Modules"`
+}
+type TerraformInitModule struct {
+	Key    string `json:"Key"`
+	Source string `json:"Source"`
+	Dir    string `json:"Dir"`
+}
+
+func (me *Environment) RunQuickInit() error {
+
+	if QUICK_INIT {
+		// pass
+	} else {
+		return nil
+	}
+
+	fmt.Println("Executing Quick Init ...")
+	err := me.WriteMainProviderFile()
+	if err != nil {
+		return err
+	}
+
+	err = me.RunTerraformInit()
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (me *Environment) WriteQuickModulesJSON() error {
+
+	if QUICK_INIT {
+		// pass
+	} else {
+		return nil
+	}
+
+	modules := []TerraformInitModule{}
+	modules = append(modules, TerraformInitModule{Key: "", Source: "", Dir: ""})
+
+	for _, module := range me.Modules {
+		moduleNameTrimmed := module.Type.Trim()
+		modules = appendModule(modules, moduleNameTrimmed)
+
+		if module.SplitPathModuleNameMap != nil {
+			for _, splitName := range module.SplitPathModuleNameMap {
+				modules = appendModule(modules, splitName)
+			}
+		}
+
+	}
+
+	tfInitModules := TerraformInitModuleList{
+		Modules: modules,
+	}
+
+	bytes, err := json.MarshalIndent(tfInitModules, "", "  ")
+	if err != nil {
+		return err
+	}
+
+	modulesDir := filepath.Join(me.OutputFolder, ".terraform", "modules")
+	os.MkdirAll(modulesDir, os.ModePerm)
+	fs := afero.NewOsFs()
+	filename := fmt.Sprint(filepath.Join(modulesDir, "modules.json"))
+	err = afero.WriteFile(fs, filename, bytes, 0664)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func readStuff(scanner *bufio.Scanner) {
+	for scanner.Scan() {
+		fmt.Println(scanner.Text())
+	}
+	if err := scanner.Err(); err != nil {
+		fmt.Fprintln(os.Stderr, "reading standard input:", err)
+	}
+}
+
+func appendModule(modules []TerraformInitModule, moduleNameTrimmed string) []TerraformInitModule {
+	modules = append(modules, TerraformInitModule{
+		Key:    moduleNameTrimmed,
+		Source: fmt.Sprintf("./modules/%s", moduleNameTrimmed),
+		Dir:    fmt.Sprintf("modules/%s", moduleNameTrimmed),
+	})
+	return modules
 }
